+++ CHANGELOG.md
@@ -4,9 +4,24 @@
 
 ** Updated: June 1st, 2025 **
 
-- Added `lazygit.nix` to theme, disable popup, etc lazygit
+- Modified zsh config
+
+```nix
+ syntaxHighlighting = {
+      enable = true;
+      highlighters = ["main" "brackets" "pattern" "regexp" "root" "line"];
+    };
+    historySubstringSearch.enable = true;
+
+    history = {
+      ignoreDups = true;
+      save = 10000;
+      size = 10000;
+    };
+```
+
+- Added `lazygit.nix` to theme, customize lazygit util
 - Added `fzf.nix` to customize fzf util
-- Added v2.3.1 to version to indicated post GA code
 - Added `waybar-ddubs-2.nix` Modified version of Jerry's waybar
 - Adjusted some colors in Jerry's waybar
 - Disabled the `df` command in the disk module. Doesn't work w/zaneyos